check to make sure the file exist before delete

+++ postinstall.js
@@ -5,5 +5,7 @@ const filesToRemove = [
 ];
 
 filesToRemove.forEach(function (filePath) {
-  fs.unlinkSync(filePath);
+  if (fs.existsSync(filePath)) {
+    fs.unlinkSync(filePath);
+  }
 });
